Adds documentation to deploy containers (#10)
The objective of Sprint 0 is to show how to get any communication between containers on a host. This is achieved by setting up a docker swarm that communicates over a common network. The communication is verified using a script docker/ping2host.sh, which makes each node ping every other node and report the results back.

Instructions on how to set up the swarm is in the README.md

Co-authored-by: freeolo <70589657+freeolo@users.noreply.github.com>