Add store/fetch and CLI
Pipes the log/stdout to a file last.log, so that the only output on the
prompt is stderr and stdin.

The prompt supports the commands put get and exit. I have tested that
nodes without a value stored on them can fetch them on command.

Known bug: Currently does not force parallelism to cap at ALPHA,
it caps forking to ALPHA, leading to an exponential increase in
running goroutines. This is not a problem on small networks though.